Implement new HTTP layer to support the new version of OpenCMIS

+++ tests/alfresco-mobile-android-unit-test-api/src/org/alfresco/mobile/android/test/AlfrescoSDKTestCase.java
@@ -347,7 +347,9 @@ public abstract class AlfrescoSDKTestCase extends InstrumentationTestCase implem
             {
                 wait(3000);
                 folder = (Folder) docfolderservice.getChildByPath(parentFolder, folderName);
-                docfolderservice.deleteNode(folder);
+                if (folder != null){
+                    docfolderservice.deleteNode(folder);
+                }
                 wait(3000);
                 folder = docfolderservice.createFolder(parentFolder, folderName, properties);
             }
